+++ examples/examples/sleep_not_co.rs
@@ -1,0 +1,7 @@
+use open_coroutine_examples::sleep_test;
+
+#[open_coroutine::main]
+fn main() {
+    sleep_test(1);
+    sleep_test(1000);
+}
