+++ examples/examples/socket_not_co.rs
