--- src/tests/mod.rs
@@ -1,8 +1,0 @@
-use super::*;
-
-mod sleep;
-
-#[test]
-fn test_link() {
-    init();
-}
